StreamlitApp/requirements.txt
streamlit
mysql-connector-python
pandas
boto3